% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tabs.R
\name{DEtoXlSX}
\alias{DEtoXlSX}
\title{Title save different expression genes to excel files}
\usage{
DEtoXlSX(Diff_res, Dir)
}
\arguments{
\item{Dir}{default dir path: "2.Results/R_outputs/Tables/"}
}
\description{
Title save different expression genes to excel files
}
