% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{fill_TPM}
\alias{fill_TPM}
\title{Title TPM normalized of raw counts}
\usage{
fill_TPM(object = NULL, version = "V3", gtf_file = NULL)
}
\arguments{
\item{gtf_file}{}
}
\description{
Title TPM normalized of raw counts
}
