% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RNASeq.R
\name{get_mRNA_diff}
\alias{get_mRNA_diff}
\title{Title get different expression results}
\usage{
get_mRNA_diff(
  dds = dds,
  group = "Condition",
  x,
  y,
  p_cutoff = 0.05,
  FC_cutoff = 1
)
}
\arguments{
\item{dds}{dds}

\item{group}{default.name = "Condition"}

\item{x}{'group1'}

\item{y}{'group1'}

\item{p_cutoff}{default: 0.05}

\item{FC_cutoff}{default: 1}
}
\description{
Title get different expression results
}
