% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{get_termList}
\alias{get_termList}
\title{Title}
\usage{
get_termList(
  geneCls,
  topN = 10,
  fun = "enrichGO",
  ont = "BP",
  OrgDb = "org.SjaponicumV4.eg.db",
  keyType = "GENEID",
  text_limits = c(8, 12)
)
}
\arguments{
\item{text_limits}{}
}
\description{
Title
}
