% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_KEGG.R
\name{getGO}
\alias{getGO}
\title{Title}
\usage{
getGO(genes, universeIDs, qCutoff = 0.05, version = "V3")
}
\arguments{
\item{version}{default: V3 , V3 -- "org.Sjaponicum.eg.db" V4 -- "org.SjaponicumV4.eg.db"}
}
\description{
Title
}
