% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{plot_complexHeatmap}
\alias{plot_complexHeatmap}
\title{Title}
\usage{
plot_complexHeatmap(
  RNASeqOBJ,
  Diff_res,
  topN = 5,
  anno_cols = c(`Up-regulated` = "#c77cff", `Down-regulated` = "#FF9900"),
  annoblock = TRUE,
  annoGO = TRUE
)
}
\arguments{
\item{annoGO}{}
}
\description{
Title
}
