% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R, R/plot_functions.R
\name{plot_corHeatmap}
\alias{plot_corHeatmap}
\title{Title}
\usage{
plot_corHeatmap(
  object = NULL,
  assay = "tpm",
  method = "pearson",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  width = 8,
  height = 8,
  show = TRUE
)

plot_corHeatmap(
  object = NULL,
  assay = "tpm",
  method = "pearson",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  width = 8,
  height = 8,
  show = TRUE
)
}
\arguments{
\item{show}{}
}
\description{
Title

Title
}
