% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Plot_corHeatmap_dds}
\alias{Plot_corHeatmap_dds}
\title{Title}
\usage{
Plot_corHeatmap_dds(dds, fileDir = "./", type = "pearson", savePlot = FALSE)
}
\arguments{
\item{type}{}
}
\description{
Title
}
