% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Plot_enhancedVolcano}
\alias{Plot_enhancedVolcano}
\title{Title}
\usage{
Plot_enhancedVolcano(
  Diff_res,
  n = 5,
  Dir = "./",
  save_Plot = FALSE,
  p_cutoff = 0.05,
  FC_cutoff = 1
)
}
\arguments{
\item{wt}{}
}
\description{
Title
}
