% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\docType{class}
\name{RNASeqOBJ-class}
\alias{RNASeqOBJ-class}
\title{Title 定义一个名为Person的S4类，它有三个属性：Assay 类 , pheno 和 outDir}
\description{
Title 定义一个名为Person的S4类，它有三个属性：Assay 类 , pheno 和 outDir
}
\section{Slots}{

\describe{
\item{\code{assay}}{Assay.}

\item{\code{pheno}}{data.frame.}

\item{\code{outDir}}{character.}

\item{\code{dds}}{DESeqDataSet.}

\item{\code{contrast_df}}{data.frame.}

\item{\code{diffres}}{list.}
}}

