% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{RNASeqOBJ}
\alias{RNASeqOBJ}
\title{Title create a RNAseq object for run RNAseq pipeline}
\usage{
RNASeqOBJ(counts, pheno = data.frame(), outDir = "./")
}
\arguments{
\item{counts}{raw counts of RNA-seq data}

\item{pheno}{a dataframe including Sample, Condition and other information}

\item{outDir}{output path}
}
\description{
Title create a RNAseq object for run RNAseq pipeline
}
