% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RNASeq.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Title A pipeline of run RNA-seq DEGs analysis and mini database of Schistosoma japonicum}
\usage{
run_DESeq2(
  count_data,
  group_a_name,
  group_a_samples,
  group_b_name,
  group_b_samples,
  pheno_file = "",
  batch = NULL
)
}
\arguments{
\item{count_data}{file path of count matrix which generated by previous RNA-seq pipeline,
eg: "2.Results/SnakePipe_outputs_test/featureCounts/counts.tsv"}

\item{group_a_name}{group a name eg:GFP}

\item{group_a_samples}{a list of samples in group A you choose, must in the colname of count_data}

\item{group_b_name}{group b name eg:dsNFY}

\item{group_b_samples}{a list of samples in group B you choose, must in the colname of count_data}

\item{pheno_file}{if group > 2, please provide pheno info in a csv file format.}

\item{batch}{if have batch effect, eg c(1,1,1,2,2,1),1~ batch1,2~ batch2}
}
\value{
a dds class for downstream analysis
}
\description{
Title A pipeline of run RNA-seq DEGs analysis and mini database of Schistosoma japonicum
}
\examples{
dds <- run_DESeq2(
  count_data = "~/Projects/RNASeq_CGW_20221014/2.Results/SnakePipe_outputs/featureCounts/counts.tsv",
  group_a_name = 'GFP',
  group_a_samples  = c("G.F1","G.F3","G.M"),
  group_b_name = 'WSS',
  group_b_samples = c("W.F1","W.F2","W.M1")
)
}
