% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_plots.R
\name{scBubbHeat}
\alias{scBubbHeat}
\title{Title Plot gene expression bubbleplot / heatmap}
\usage{
scBubbHeat(
  shinycell_prefix,
  inp,
  inpGrp,
  inpPlt = "Bubbleplot",
  inpsub1 = NULL,
  inpScl = T,
  inpRow = T,
  inpCol = T,
  inpcols = "Blue-Yellow-Red",
  inpfsz = "Small",
  save = FALSE
)
}
\arguments{
\item{save}{}
}
\description{
Title Plot gene expression bubbleplot / heatmap
}
