% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SJDB-package.R
\docType{package}
\name{SJDB-package}
\alias{SJDB}
\alias{SJDB-package}
\title{SJDB: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: ShaoYun Cheng \email{shawyunchan@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
