% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{plot_corHeatmap}
\alias{plot_corHeatmap}
\title{Title plot correlation coefficient heatmap of samples}
\usage{
plot_corHeatmap(
  object = NULL,
  assay = "tpm",
  method = "pearson",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  show = TRUE
)
}
\arguments{
\item{show}{}
}
\description{
Title plot correlation coefficient heatmap of samples
}
