% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Title plot_PCA using one of assay in RNAseq object}
\usage{
plot_PCA(object = NULL, assay = "tpm", show = TRUE)
}
\arguments{
\item{show}{}
}
\description{
Title plot_PCA using one of assay in RNAseq object
}
