% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sjV4.R, R/plot_functions.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Title plot_PCA using one of assay in RNAseq object}
\usage{
plot_PCA(
  dds,
  save_Plot = FALSE,
  file_Dir = "./",
  file_Name = "All_Samples_PCA_removeBatch.pdf"
)

plot_PCA(
  dds,
  save_Plot = FALSE,
  file_Dir = "./",
  file_Name = "All_Samples_PCA_removeBatch.pdf"
)
}
\arguments{
\item{show}{}
}
\description{
Title plot_PCA using one of assay in RNAseq object

Title plot PCA
}
