% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RNASeq.R
\name{run_RNASeq}
\alias{run_RNASeq}
\title{Title}
\usage{
run_RNASeq(
  count_data,
  group_a_name,
  group_a_samples,
  group_b_name,
  group_b_samples,
  pheno_file = "",
  batch = NULL,
  p_cutoff = 0.05,
  FC_cutoff = 1,
  save_Plot,
  file_Dir
)
}
\arguments{
\item{file_Dir}{}
}
\description{
Title
}
