% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_plots.R
\docType{class}
\name{ShinyCell_prefix-class}
\alias{ShinyCell_prefix-class}
\alias{ShinyCell_prefix}
\title{Title a S4 object of shinycell_prefix}
\description{
Title a S4 object of shinycell_prefix
}
\section{Slots}{

\describe{
\item{\code{conf}}{data.table.}

\item{\code{def}}{list.}

\item{\code{gene}}{integer.}

\item{\code{meta}}{data.table.}

\item{\code{cList}}{list.}

\item{\code{sList}}{numeric.}

\item{\code{hdf5}}{character.}
}}

